fix: choosing available taxonomic group types
